+++ Examples/FXPageControlExample/AppDelegate.h
@@ -1,0 +1,18 @@
+//
+//  CustomPageControlExampleAppDelegate.h
+//  CustomPageControlExample
+//
+//  Created by Nick Lockwood on 03/04/2011.
+//  Copyright 2011 __MyCompanyName__. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@class ViewController;
+
+@interface AppDelegate : NSObject <UIApplicationDelegate>
+
+@property (nonatomic, strong) IBOutlet UIWindow *window;
+@property (nonatomic, strong) IBOutlet ViewController *viewController;
+
+@end