+++ Examples/FXPageControlExample/AppDelegate.m
@@ -1,0 +1,22 @@
+//
+//  CustomPageControlExampleAppDelegate.m
+//  CustomPageControlExample
+//
+//  Created by Nick Lockwood on 03/04/2011.
+//  Copyright 2011 __MyCompanyName__. All rights reserved.
+//
+
+#import "AppDelegate.h"
+#import "ViewController.h"
+
+
+@implementation AppDelegate
+
+- (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
+{
+    self.window.rootViewController = self.viewController;
+    [self.window makeKeyAndVisible];
+    return YES;
+}
+
+@end