+++ RELEASE NOTES.md
@@ -1,3 +1,11 @@
+Version 1.2
+
+- Renamed to FXPageControl
+- FXPageControl now requires ARC. See README for details
+- FXPageControl is now a subclass of UIPageControl
+- Added support for custom dot images
+- Renamed wrap property to wrapEnabled
+
 Version 1.1.1
 
 - View now redraws itself automatically if bounds are changed